refactor: removed unused code
