+++ src/components/Wishlist/Wishlist.js
@@ -2,16 +2,11 @@ import React from "react";
 import "../../css/wishlist.css";
 import { useMainContext } from "../../context/context";
 import { Wishcard, Wishlistempty } from "../index";
-import { useWishlistData, useCartData } from "../../hooks/index";
 import { clearWishlist } from "../../api/wishlist/clearWishlist";
-import { useAuth } from "../../context/auth-context";
 import { PuffLoader } from "../Loader";
 
 export const Wishlist = () => {
   const { loader, wishList, dispatch } = useMainContext();
-  const { token } = useAuth();
-  useCartData();
-  useWishlistData();
 
   if (wishList.length === 0) {
     return (
@@ -25,7 +20,7 @@ export const Wishlist = () => {
       <div className="wishlist-section">
         <button
           className="btn btn--round btn-secondary"
-          onClick={() => clearWishlist(dispatch, token)}>
+          onClick={() => clearWishlist(dispatch)}>
           Clear Wishlist
         </button>
         <div className="wish-container">
